The code now compiles and runs on Fedora 9 (64 bit, compiled with gcc 4.3). I needed to make a few modifications to fix some compile bugs.
I've also switched mpinode to use Q_GLOBAL_STATIC for the QMutex pointers, in an attempt to 
fix the QMutex errors on shutdown. I can't test it though, as I deadlock on linux, and 
have lseek64 symbol errors on my mac again!

+++ corelib/branches/users/chryswoods/src/libs/SireMM/ljparameterdb.cpp
@@ -230,7 +230,7 @@ quint32 LJParameterDB::LJParameterDBData::addLJParameter(const LJParameter &ljpa
 LJParameter LJParameterDB::LJParameterDBData::_locked_getLJParameter(quint32 id) const
 {
     //does the parameter exist already in the database?
-    if (id > ljparams_by_idx.count())
+    if (id > quint32(ljparams_by_idx.count()))
         throw SireFF::missing_parameter( QObject::tr(
             "Could not find the LJ parameter with ID %1 in the global "
             "LJ parameter database.").arg(id), CODELOC );
